nsaid/pyproject.toml
[tool.poetry]
name = "nsaid"
version = "0.1.0"
description = ""
authors = ["Daniel Wagner <me@dmwit.com>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.4"
tensorboard = "^2.2"
setuptools = "^41.0"
pillow = "^7.1.2"
torchvision = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
